+++ gear/participant_transfer/src/python/participant_transfer_app/transfer.py
@@ -353,17 +353,20 @@ class TransferProcessor:
             return False
 
         # update the record in previous center's enrollment project
-        prev_enroll_project = prev_center.find_project(self.__enroll_project.label)
-        if (
-            not prev_enroll_project
-            and self.__enroll_project.label != DefaultValues.ENRL_PRJ_LABEL
-        ):
-            prev_enroll_project = prev_center.find_project(DefaultValues.ENRL_PRJ_LABEL)
-        # TODO: Fix looking up prev enrollment project (check pipeline ADCID)
+        prev_enroll_project = None
+        project_matches = prev_center.get_matching_projects(
+            prefix=DefaultValues.ENRL_PRJ_LABEL
+        )
+        for project in project_matches:
+            if project.get_pipeline_adcid() == self.__transfer_record.previous_adcid:
+                prev_enroll_project = project
+                break
+
         if not prev_enroll_project:
             message = (
-                "Failed to find enrollment project in previous center "
-                f"{prev_center.label}, ADCID: {prev_center.adcid}"
+                "Failed to find enrollment project in "
+                f"previous center {prev_center.label} "
+                f"with ADCID: {self.__transfer_record.previous_adcid}"
             )
             log.warning(message)
             self.__warnings.append(message)