+++ docs/identifier_lookup/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 All notable changes to this gear are documented in this file.
 
-## Unreleased
+## 2.0.0
+* Changes to use the ADCID from the parent project custom info for the pipeline ADCID. Fails if the ADCID is not set.
 * Updates to use new identifier lambda functions with restructured identifiers database
 * Updates error metadata to include date and naccid
 * Make `form_configs_file` only required for `nacc` direction